Change tasks inheritance.
Now KtlintCheck and KtlintFormat tasks extends base BaseKtlintCheckTask.
